agent/requirements.txt
# Core LiveKit packages
livekit-agents>=1.0.11
livekit-plugins-silero>=0.7.5
livekit-plugins-noise-cancellation>=0.2.5
livekit-plugins-turn-detector>=1.0
livekit-plugins-google>=0.11.0
livekit-plugins-groq>=0.7.0

# Python utilities
python-dotenv~=1.0
asyncpg~=0.30.0
aiofile~=3.8.8
fastapi~=0.115
uvicorn[standard]~=0.32
httpx>=0.28.1
PyJWT~=2.8.0